# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API E-commerce",
        "description": "API E-commerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get list of orders",
                "operationId": "00d3856d2d3d93124e0cc021d0cb5b79",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Create a orders",
                "operationId": "79da0a1adcc617490783dfd61927ff56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/{order}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get a order by ID",
                "operationId": "38713af501dbd66af5a66bd568a49bc2",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "orders"
                ],
                "summary": "Delete a order",
                "operationId": "4baa2e826530c2663217ed85efd8e3e3",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order deleted successfully"
                    }
                }
            },
            "patch": {
                "tags": [
                    "orders"
                ],
                "summary": "Update an order",
                "operationId": "25408827628495f84fbaecb37dd89c84",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get list of products",
                "operationId": "1ce27f90dd8c39cea9520bf45d58c7c2",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Create a product",
                "operationId": "acaa0d3fdc74fd4b534e70fc016cf7f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{product}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get a product by ID",
                "operationId": "9d6564b9647b6b46ca1d3eb86c51643e",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "operationId": "dc780252f855e97451083ce643a83174",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    }
                }
            },
            "patch": {
                "tags": [
                    "products"
                ],
                "summary": "Update a product",
                "operationId": "780ef5766bbb62de4378b38eb7cdfbda",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "description": "Title of the product",
                        "type": "string",
                        "example": "Product 1"
                    },
                    "description": {
                        "description": "Description of the product",
                        "type": "string",
                        "example": "Description of the product"
                    },
                    "price": {
                        "description": "Price of the product",
                        "type": "number",
                        "example": 1000
                    },
                    "user_id": {
                        "description": "ID of the user who created the product",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreProductRequest": {
                "required": [
                    "title",
                    "description",
                    "price",
                    "user_id"
                ],
                "properties": {
                    "title": {
                        "description": "Title of the product",
                        "type": "string",
                        "example": "Product 1"
                    },
                    "description": {
                        "description": "Description of the product",
                        "type": "string",
                        "example": "Description of the product"
                    },
                    "price": {
                        "description": "Price of the product",
                        "type": "number",
                        "example": 1000
                    },
                    "user_id": {
                        "description": "ID of the user who created the product",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UpdateProductRequest": {
                "properties": {
                    "title": {
                        "description": "Title of the product",
                        "type": "string",
                        "example": "Updated Product"
                    },
                    "image": {
                        "description": "Image of the product",
                        "type": "string",
                        "example": "<url>"
                    },
                    "description": {
                        "description": "Description of the product",
                        "type": "string",
                        "example": "Updated description"
                    },
                    "price": {
                        "description": "Price of the product",
                        "type": "number",
                        "example": 1500
                    },
                    "user_id": {
                        "description": "ID of the user who created the product",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "OrderResponse": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object"
            },
            "ProductOrderRequest": {
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 4
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "CreateOrderRequest": {
                "properties": {
                    "user_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string",
                        "example": "pendiente"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductOrderRequest"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateOrderRequest": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "pendiente"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductOrderRequest"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "orders",
            "description": "orders"
        },
        {
            "name": "products",
            "description": "products"
        }
    ]
}